Path C: split composeApp into shared + per-platform app modules
JetBrains' kotlin-tooling-agp9-migration skill describes "Path C: Full
Restructure" as the recommended layout under AGP 9 — one shared KMP
library plus a dedicated app module per platform. Apply it here so we
can drop the android.builtInKotlin=false / android.newDsl=false
opt-out flags that the previous commit needed.

- composeApp → shared (KMP library only): commonMain App.kt + theme +
  shared DI interfaces, iosMain MainViewController + iOS DI, iOS
  framework export. Uses the new com.android.kotlin.multiplatform.library
  plugin (via KotlinMultiplatformConventionPlugin) so the android target
  is still produced for androidApp to depend on.
- androidApp/ (new com.android.application module): MainActivity,
  RijksmuseumApplication, AndroidAppComponent + AndroidUiComponent,
  manifest, launcher icons, signing config. Built-in Kotlin (no
  kotlin-android plugin) per AGP 9 default.
- desktopApp/ (new KMP module, jvm-only): main.kt + Desktop DI +
  compose.desktop config.
- webApp/ (new KMP module, wasmJs-only): main.kt + Web DI + index.html.

Supporting changes:
- shared exposes UI deps via api() so each app module's KSP pass can
  see all @CircuitInject / @ContributesBinding classes when generating
  the merged DI graph.
- KotlinAndroid.kt: namespace now uses the full gradle path joined by
  '.', so ':shared' becomes "com.tewelde.rijksmuseum.shared" (not the
  base) — AGP 9's android.uniquePackageNames=true rejects the old
  collision with androidApp's "com.tewelde.rijksmuseum".
- settings.gradle.kts: drop :composeApp, include :shared :androidApp
  :desktopApp :webApp.
- iosApp/iosApp.xcodeproj: gradle hook now invokes
  :shared:embedAndSignAppleFrameworkForXcode.
- webApp index.html: <script src="webApp.js"> (was composeApp.js +
  skiko.js, which the new Compose Multiplatform bundles internally).
- gradle.properties: drop android.builtInKotlin=false /
  android.newDsl=false — no longer needed.

Verified: ./gradlew check passes; runtime on Android (Pixel 7,
PID 3754), iOS simulator (iPhone 17, screenshot), and wasm (headless
Chrome via CDP with software WebGL) all load the Arts pager and the
"Before, Behind, Between, Above, Below" IIIF image.
