+++ androidApp/build.gradle.kts
@@ -1,0 +1,80 @@
+import java.util.Properties
+
+plugins {
+    alias(libs.plugins.androidApplication)
+    alias(libs.plugins.jetbrainsCompose)
+    alias(libs.plugins.compose.compiler)
+    alias(libs.plugins.ksp)
+}
+
+val secretKeyProperties: Properties by lazy {
+    val secretKeyPropertiesFile = rootProject.file("secrets.properties")
+    Properties().apply { secretKeyPropertiesFile.inputStream().use { secret -> load(secret) } }
+}
+
+android {
+    namespace = "com.tewelde.rijksmuseum"
+    compileSdk = libs.versions.android.compileSdk.get().toInt()
+
+    defaultConfig {
+        applicationId = "com.tewelde.rijksmuseum"
+        minSdk = libs.versions.android.minSdk.get().toInt()
+        targetSdk = libs.versions.android.targetSdk.get().toInt()
+        versionCode = 5
+        versionName = "1.0.5"
+    }
+    packaging {
+        resources {
+            excludes += "/META-INF/{AL2.0,LGPL2.1}"
+        }
+    }
+    signingConfigs {
+        create("release") {
+            storeFile = file("$rootDir/keystore/rijksmuseum.jks")
+            storePassword = "${secretKeyProperties["rijksmuseum.keystore.password"]}"
+            keyAlias = "${secretKeyProperties["rijksmuseum.key.alias"]}"
+            keyPassword = "${secretKeyProperties["rijksmuseum.key.password"]}"
+        }
+    }
+    buildTypes {
+        debug {
+            isMinifyEnabled = false
+            isDebuggable = true
+            applicationIdSuffix = ".debug"
+        }
+        release {
+            isDebuggable = false
+            isMinifyEnabled = true
+            isShrinkResources = true
+            proguardFiles(
+                getDefaultProguardFile("proguard-android-optimize.txt"),
+                "proguard-rules.pro"
+            )
+            signingConfig = signingConfigs.getByName("release")
+        }
+    }
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_21
+        targetCompatibility = JavaVersion.VERSION_21
+    }
+    buildFeatures {
+        compose = true
+    }
+}
+
+dependencies {
+    implementation(projects.shared)
+    implementation(projects.core.common)
+    implementation(projects.core.navigation)
+    implementation(projects.core.permissions)
+
+    implementation(libs.androidx.activity.compose)
+    implementation(libs.kotlinx.coroutines.android)
+    implementation(libs.bundles.kotlinInjectAnvil)
+    implementation(libs.circuit.foundation)
+
+    ksp(libs.kotlinInject.compiler)
+    ksp(libs.kotlinInject.anvil.compiler)
+
+    debugImplementation(compose.uiTooling)
+}
