+++ androidApp/src/main/res/mipmap-anydpi-v26/rijksmuseum_launcher_round.xml
