+++ androidApp/src/main/res/values/rijksmuseum_launcher_background.xml
