--- composeApp/build.gradle.kts
@@ -1,158 +1,0 @@
-import com.tewelde.rijksmuseum.addKspDependencyForAllTargets
-import org.jetbrains.compose.desktop.application.dsl.TargetFormat
-import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
-import java.util.Properties
-
-plugins {
-    alias(libs.plugins.kotlinMultiplatform)
-    alias(libs.plugins.androidApplication)
-    alias(libs.plugins.jetbrainsCompose)
-    alias(libs.plugins.compose.compiler)
-    alias(libs.plugins.rijksmuseum.featureMultiplatform)
-}
-
-val secretKeyProperties: Properties by lazy {
-    val secretKeyPropertiesFile = rootProject.file("secrets.properties")
-    Properties().apply { secretKeyPropertiesFile.inputStream().use { secret -> load(secret) } }
-}
-
-kotlin {
-    jvmToolchain(libs.versions.jvmTarget.get().toInt())
-
-    androidTarget {}
-
-    jvm()
-
-    @OptIn(ExperimentalWasmDsl::class)
-    wasmJs {
-        browser()
-        binaries.executable()
-    }
-
-    listOf(
-        iosArm64(),
-        iosSimulatorArm64()
-    ).forEach { iosTarget ->
-        iosTarget.binaries.framework {
-            baseName = "ComposeApp"
-            isStatic = true
-            optimized = true
-        }
-    }
-
-    sourceSets {
-        androidMain.dependencies {
-            implementation(libs.androidx.activity.compose)
-            implementation(libs.ktor.client.okhttp)
-            implementation(libs.kotlinx.coroutines.android)
-        }
-
-        commonMain.dependencies {
-            implementation(projects.feature.arts)
-            implementation(projects.feature.detail)
-            implementation(projects.core.navigation)
-            implementation(projects.core.permissions)
-
-            implementation(compose.material3)
-            implementation(compose.components.resources)
-            implementation(compose.components.uiToolingPreview)
-
-            implementation(libs.ktor.client.core)
-            implementation(libs.coil.compose)
-            implementation(libs.coil.network.ktor)
-
-            implementation(libs.filekit.core)
-
-            implementation(libs.bundles.circuit)
-            implementation(libs.bundles.kotlinInjectAnvil)
-        }
-
-        jvmMain.dependencies {
-            implementation(compose.desktop.currentOs)
-            implementation(libs.kotlinx.coroutines.swing)
-            implementation(libs.ktor.client.java)
-        }
-
-        iosMain.dependencies {
-            implementation(libs.ktor.client.darwin)
-        }
-    }
-}
-
-addKspDependencyForAllTargets(libs.circuit.codegen)
-addKspDependencyForAllTargets(libs.kotlinInject.compiler)
-addKspDependencyForAllTargets(libs.kotlinInject.anvil.compiler)
-
-android {
-    namespace = "com.tewelde.rijksmuseum"
-    compileSdk = libs.versions.android.compileSdk.get().toInt()
-
-    sourceSets["main"].manifest.srcFile("src/androidMain/AndroidManifest.xml")
-    sourceSets["main"].res.srcDirs("src/androidMain/res")
-    sourceSets["main"].resources.srcDirs("src/commonMain/resources")
-
-    defaultConfig {
-        applicationId = "com.tewelde.rijksmuseum"
-        minSdk = libs.versions.android.minSdk.get().toInt()
-        targetSdk = libs.versions.android.targetSdk.get().toInt()
-        versionCode = 5
-        versionName = "1.0.5"
-    }
-    packaging {
-        resources {
-            excludes += "/META-INF/{AL2.0,LGPL2.1}"
-        }
-    }
-    signingConfigs {
-        create("release") {
-            storeFile = file("rijksmuseum.jks")
-            storePassword = "rijksmuseum"
-            keyAlias = "rijksmuseum"
-            keyPassword = "rijksmuseum"
-            storeFile = file("$rootDir/keystore/rijksmuseum.jks")
-            storePassword = "${secretKeyProperties["rijksmuseum.keystore.password"]}"
-            keyAlias = "${secretKeyProperties["rijksmuseum.key.alias"]}"
-            keyPassword = "${secretKeyProperties["rijksmuseum.key.password"]}"
-        }
-    }
-    buildTypes {
-        debug {
-            isMinifyEnabled = false
-            isDebuggable = true
-            applicationIdSuffix = ".debug"
-        }
-
-        release {
-            isDebuggable = false
-            isMinifyEnabled = true
-            isShrinkResources = true
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-            signingConfig = signingConfigs.getByName("release")
-        }
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_21
-        targetCompatibility = JavaVersion.VERSION_21
-    }
-    buildFeatures {
-        compose = true
-    }
-    dependencies {
-        debugImplementation(compose.uiTooling)
-    }
-}
-
-compose.desktop {
-    application {
-        mainClass = "MainKt"
-
-        nativeDistributions {
-            targetFormats(TargetFormat.Dmg, TargetFormat.Msi, TargetFormat.Deb)
-            packageName = "com.tewelde.rijksmuseum"
-            packageVersion = "1.0.0"
-        }
-    }
-}
