+++ gradle.properties
@@ -11,11 +11,6 @@ kotlin.code.style=official
 #Android
 android.useAndroidX=true
 android.nonTransitiveRClass=true
-# AGP 9.0 made the android.application/library plugins incompatible with kotlin-multiplatform.
-# Until we migrate library modules to com.android.kotlin.multiplatform.library, opt out of the
-# new built-in Kotlin support and the new DSL.
-android.builtInKotlin=false
-android.newDsl=false
 
 #Compose
 org.jetbrains.compose.experimental.wasm.enabled=true
