+++ webApp/src/wasmJsMain/resources/index.html
@@ -13,9 +13,8 @@
             overflow: hidden;
         }
     </style>
-    <script src="skiko.js"></script>
+    <script type="application/javascript" src="webApp.js"></script>
 </head>
 <body>
-<script src="composeApp.js"></script>
 </body>
 </html>
